+++ cargo-near/src/abi/mod.rs
@@ -1,10 +1,9 @@
+use super::AbiCommand;
 use crate::cargo::{manifest::CargoManifestPath, metadata::CrateMetadata};
 use crate::util;
 use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
-
-use super::AbiCommand;
 
 mod generation;
