+++ cargo-near/src/lib.rs
@@ -1,6 +1,5 @@
-use std::path::PathBuf;
-
 use clap::{AppSettings, Args, Parser, Subcommand};
+use std::path::PathBuf;
 
 mod abi;
 mod cargo;
